新增：消息盒子功能开发

+++ Micro-WebIM/src/main/java/com/micro/im/ws/dto/AddFriendMessage.java
@@ -26,7 +26,7 @@ public class AddFriendMessage {
     /**
      * 好友分组ID
      */
-    private Integer groupid;
+    private Long groupid;
     /**
      * 好友ID
      */
